Adjustments to navigation along by adding it to a component
The nav component still needs a better refactor. BEM writing style doesn't
seem accurate as it is right now. All of this code should be more DRY, and not
only that, but we may be able to take some styles from here and make it into
other components, like .logo and .user-profile.